Email Classification(P1)/requirements.txt
streamlit==1.20.0
scikit-learn==1.2.1
pandas==1.5.3
numpy==1.24.2
Pillow==9.4.0
